Make sure grid is not outside of document.
Stroke width is considered when calculating hex
sizes. The grid is also shifted half the width
of the stroke to make sure the stroke stays
inside of the page area.

The implementation is pretty crude at the moment.
There should be a better way to do this.
But it is a bit late here in this timezone.